+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Incompatibility for users without cURL extension but with enabled `allow_url_fopen` - PR [#139](https://github.com/lavary/crunz/pull/139)
+by [@PabloKowalczyk](https://github.com/PabloKowalczyk)
+
 ## 1.10.0 - 2018-09-22
 
 ### Fixed
@@ -18,7 +23,7 @@ by [@PabloKowalczyk](https://github.com/PabloKowalczyk)
 ### Removed
 
 - Remove `guzzlehttp/guzzle` dependency and use in-house `CurlHttpClient` -
-PR [#132](https://github.com/lavary/crunz/pull/136)
+PR [#136](https://github.com/lavary/crunz/pull/136)
 by [@PabloKowalczyk](https://github.com/PabloKowalczyk)
 
 ## 1.9.0 - 2018-08-18